converted file structure to TS, added tsconfig, updated Jest
