+++ server/controllers/dbController.ts
@@ -2,4 +2,4 @@ const db = require('../models/dbModel.js');
 
 const dbController = {};
 
-module.exports = dbController;
+export default dbController;
