+++ server/models/dbModel.ts
@@ -6,8 +6,8 @@ import "dotenv/config.js";
 //import { MONGO_URI } from '../secret';
 //import secret from '../secret'
 mongoose.connect(process.env.MONGO_URI, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
+  // useNewUrlParser: true,
+  // useUnifiedTopology: true,
   dbName: 'EkhoMS'
 })
   .then(() => console.log('Connected to Mongo DB.'))
